requirements.txt
# Core dependencies - relaxed for Python 3.14 compatibility
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1

# Scheduling and background jobs
apscheduler>=3.10.4
celery>=5.3.6
redis>=5.0.1

# AI/LLM integration
openai>=1.12.0
anthropic>=0.18.0

# Utilities
pydantic[email]
python-dateutil>=2.8.2
jinja2>=3.1.3
pydantic-yaml>=1.2.1
loguru>=0.7.2

# Frontend (optional, simple dashboard)
python-multipart>=0.0.6

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.5
